# feat(database): implement enhanced connection pool with ML-driven optimization (BJO-134)