… .github/ISSUE_TEMPLATE/database-optimization.md
… .github/ISSUE_TEMPLATE/documentation-improvement.md
… .github/ISSUE_TEMPLATE/ml-model-improvement.md
… .github/ISSUE_TEMPLATE/performance-regression.md
… COMMUNITY-GUIDELINES.md
… CONTRIBUTING.md
… README-MONITORING.md
… README.md
… RESEARCH-COLLABORATION.md
… ROADMAP.md
… SECURITY-REVIEW.md
… TODO.md
… docs/developers/api-reference.md
… docs/developers/architecture.md
… docs/developers/benchmarking-and-performance.md
… docs/developers/configuration.md
… docs/developers/database-connection-pool-technical-reference.md
… docs/developers/database-optimization-api.yaml
… docs/developers/enhanced-architecture-diagrams.md
… docs/developers/getting-started.md
… docs/developers/ml-model-documentation.md
… docs/developers/performance-benchmarking-methodology.md
… docs/operators/configuration.md
… docs/operators/database-connection-pool-performance.md
… docs/operators/deployment.md
… docs/operators/enhanced-security-architecture.md
… docs/operators/monitoring.md
… docs/operators/operations.md
… docs/operators/security.md
… docs/users/examples-and-recipes.md